Move get/convertBundle to OutputAreaModel
Fits better in model.
